+++ CHANGELOG.md
@@ -9,6 +9,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - HPO gene list download now has option for clinical and non-clinical genes
 ### Fixed
 - Show other causative once, even if several events point to it
+- Filtering variants by mitochondrial chromosome for cases with genome build=38
 - HPO gene search button triggers any warnings for clinical / non-existing genes also on first search
 ### Changed